argoslabs/web/selenium/requirements.txt
# pip dependent packages

# default for alabs
alabs.common

# for selenium
alabslib.selenium
